+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class UserDto {
-    private int id;
-    private String name;            // — имя или логин пользователя;
-    private String email;           // — адрес электронной почты (учтите, что два пользователя не могут иметь одинаковый адрес электронной почты).
+    private long id;
+    private String name;
+    private String email;
 }
